Cross model relation support
Cross model relations are anonymized, so publishing authorized
unit names on the relation will not work. Instead, publish
the list of autorized subnets which clients can match with
their egress-subnets to see if they have been authorized.

allowed-units is deprecated, but will remain supported for the
forseeable future for single model deployments.